+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - optimize td render
 * Bugfix
   - fix: when drag fixed right column, and the column is fixing, and drag the th right resize handle, the column width changed not correct.
+  - fix: `minWidth` `maxWidth` not work when `props.colResizable` is true.
 
 ## 0.9.1
 * Feature